Rename BaseClient to QueryBuilder
