+++ .vscode/star-rail-api.code-snippets
@@ -3,14 +3,14 @@
     "scope": "typescript",
     "prefix": "starRailClient",
     "body": [
-      "import BaseClient, { ClientOptions } from '@/BaseClient';",
+      "import QueryBuilder, { ClientOptions } from '@/QueryBuilder';",
       "import { Resources } from '@/enum';",
       "import { ${1:RESOURCE_TYPE} } from '@/types';",
       "",
       "// Replace FieldClient with the desired client",
       "import { FieldClient } from '@/clients/FieldClient';",
       "",
-      "export class $TM_FILENAME_BASE extends BaseClient<${1:RESOURCE_TYPE}> {",
+      "export class $TM_FILENAME_BASE extends QueryBuilder<${1:RESOURCE_TYPE}> {",
       "  // Example auxiliary client to fetch reference data",
       "  private fieldClient?: FieldClient;",
       "",
