+++ src/clients/AvatarsClient.ts
@@ -1,8 +1,8 @@
-import BaseClient, { ClientOptions } from '@/BaseClient';
+import QueryBuilder, { ClientOptions } from '@/QueryBuilder';
 import { Resources } from '@/enum';
 import { Avatar } from '@/types';
 
-export class AvatarsClient extends BaseClient<Avatar> {
+export class AvatarsClient extends QueryBuilder<Avatar> {
   private includeImagePaths: boolean = false;
 
   constructor(options?: ClientOptions) {
